Integração de SQLAlchemy completa

+++ models.py
@@ -3,4 +3,4 @@ from pydantic import BaseModel
 class Produto(BaseModel):
     id: int
     nome: str
-    preco : int
+    preco : float
